Revert "Refactor leetcode command to use Cog design pattern"
