+++ bot/setup.py
@@ -4,14 +4,12 @@ from bot.core import bot
 from bot.events import on_member_join, on_voice_state_update, on_ready, on_disconnect, on_message
 from bot.tasks import daily_streak_update
 from bot.commands import CommandsCog
-from cogs.leetcode import LeetCodeCog
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
 
 async def setup_commands(bot):
     await bot.add_cog(CommandsCog(bot))
-    await bot.add_cog(LeetCodeCog(bot))
     logging.info("Commands have been set up")
 
 def setup_events(bot):
