--- requires/dev.txt
@@ -1,2 +1,0 @@
-pytest
-pytest-asyncio
